Refactor project structure, move validation and null-check from storage to service.
