+++ src/main/java/ru/practicum/shareit/item/interfaces/ItemService.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.item.service;
+package ru.practicum.shareit.item.interfaces;
 
 import ru.practicum.shareit.item.dto.ItemDto;
