+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -2,13 +2,13 @@ package ru.practicum.shareit.user;
 
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.shareit.user.dto.UserDto;
-import ru.practicum.shareit.user.service.UserService;
+import ru.practicum.shareit.user.interfaces.UserService;
 
 import javax.validation.Valid;
 import java.util.List;
 
 /**
- * TODO Sprint add-controllers.
+ * User controller.
  */
 @RestController
 @RequestMapping(path = "/users")
