+++ src/main/java/ru/practicum/shareit/user/UserMapper.java
@@ -1,8 +1,10 @@
-package ru.practicum.shareit.user.mapper;
+package ru.practicum.shareit.user;
 
 import ru.practicum.shareit.user.dto.UserDto;
-import ru.practicum.shareit.user.model.User;
 
+/**
+ * User mapper.
+ */
 public class UserMapper {
 
     public static UserDto userToDto(User user) {
