+++ src/main/java/ru/practicum/shareit/user/interfaces/UserService.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.user.service;
+package ru.practicum.shareit.user.interfaces;
 
 import ru.practicum.shareit.user.dto.UserDto;
