+++ src/main/java/ru/practicum/shareit/user/interfaces/UserStorage.java
@@ -1,6 +1,6 @@
-package ru.practicum.shareit.user.storage;
+package ru.practicum.shareit.user.interfaces;
 
-import ru.practicum.shareit.user.model.User;
+import ru.practicum.shareit.user.User;
 
 import java.util.List;
